Added a file with vertex formats structures and input descriptions
Added a file to gather vertex formats structures and input descriptions for easier management.
Fixed a bug in the release where a fence was not created due to expression placement inside assert macro.
Added DirectX to the global namespace.
Additionally, current back buffer index is cached so it will be shown in PIX's events only once.